Extend AbilityAssigner with user info (and rename to AuthContext)
